Updated Board Functionality
The board is now testing with a 6x6. I have made it so it generates a board and places a cell in each space in the grid. So far it is only able to comprehend whether the cell is a bomb or not and can return whether you lose or not.

Will be working on checking how to display surrounding bomb number.